LayoutTests/fast/viewport/ios/shrink-to-fit-content-temporary-overflow-expected.txt

This test verifies that a temporary change in content width does not cause the viewport width to permanently expand to try and accomodate the content. To test manually, load the page and check that the box below reads 'PASS'. This test is only intended to run on devices with less than 1200px screen width.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS minScale is 1
PASS successfullyParsed is true

TEST COMPLETE
